+++ src/SDF/BilletterieBundle/Resources/config/routing/billetterie/admin.yml
@@ -1,11 +1,11 @@
 sdf_billetterie_administration:
     path:     /
     defaults:
-        _controller: SDFBilletterieBundle:Admin/Pages:home
+        _controller: SDFBilletterieBundle:Pages/Admin/Pages:home
     methods: GET
 
 sdf_billetterie_administration_stats:
     path:     /stats
     defaults:
-        _controller: SDFBilletterieBundle:Admin/Pages:stats
+        _controller: SDFBilletterieBundle:Pages/Admin/Pages:stats
     methods: GET
