+++ src/SDF/BilletterieBundle/Resources/config/routing/billetterie/tickets.yml
@@ -1,23 +1,23 @@
 sdf_billetterie_accessBillet:
     path:     /{id}
-    defaults: { _controller: SDFBilletterieBundle:Ticketing:accessBillet }
+    defaults: { _controller: SDFBilletterieBundle:Pages/Ticketing:accessBillet }
 
 sdf_billetterie_paramBillet:
     path:     /{id}/param
-    defaults: { _controller: SDFBilletterieBundle:Ticketing:paramBillet }
+    defaults: { _controller: SDFBilletterieBundle:Pages/Ticketing:paramBillet }
 
 sdf_billetterie_changeParamBillet:
     path:     /{id}/change
-    defaults: { _controller: SDFBilletterieBundle:Ticketing:changedParamBillet }
+    defaults: { _controller: SDFBilletterieBundle:Pages/Ticketing:changedParamBillet }
 
 sdf_billetterie_ticket_edit:
     path:     /{id}/edit
     defaults:
-        _controller: SDFBilletterieBundle:Ticketing:editTicket
+        _controller: SDFBilletterieBundle:Pages/Ticketing:editTicket
     methods: GET
 
 sdf_billetterie_ticket_update:
     path:     /{id}/edit
     defaults:
-        _controller: SDFBilletterieBundle:Ticketing:updateTicket
+        _controller: SDFBilletterieBundle:Pages/Ticketing:updateTicket
     methods: POST
