--- src/SDF/BilletterieBundle/Resources/views/Pages/Ticketing/list.html.twig
@@ -1,91 +1,0 @@
-{% extends 'SDFBilletterieBundle:Pages/Ticketing:layout.html.twig' %}
-
-{% block app_hero_content %}
-	<h1>Billetterie de la Soirée des Finaux</h1>
-{% endblock %}
-
-{% block app_content %}
-	{% if unvalidTicket %}
-		<div class='alert alert-danger' role='alert'>
-			<strong>Un de vos billets n'a pas encore été traité par PayUTC.</strong>
-			<p>Cliquez <a href='{{ path('sdf_billetterie_relancerTransaction', { 'id' : unvalidTicket.id }) }}'>ici pour reprendre la transaction</a>, ou <a href='{{ path('sdf_billetterie_annulerBilletInvalide', {'id' : unvalidTicket.id}) }}'>ici pour l'annuler</a>.</p>
-		</div>
-	{% endif %}
-
-	{% if boughtTickets|length == 0 %}
-		<h2>Aucun billet acheté.</h2>
-	{% else %}
-		<!-- Liste des billets achetés -->
-		<div class='panel panel-default'>
-			<!-- Default panel contents -->
-			<div class='panel-heading'>
-				<h3>Billets achetés</h3>
-			</div>
-			<div class='panel-body'>
-				<!-- Table -->
-				<table class='table'>
-					<thead>
-						<tr>
-							<th>Nom</th>
-							<th>Type de billet</th>
-							<th>Navette</th>
-							<th class='text-right'>Actions</th>
-						</tr>
-					</thead>
-					<tbody>
-						{% for ticket in boughtTickets %}
-							<tr>
-								<td>{{ ticket.prenom ~ ' ' ~ ticket.nom }}</td>
-								<td>{{ ticket.tarif.nomTarif }}</td>
-								<td>{% if ticket.navette %}Navette de {{ ticket.navette.trajet.lieuDepart }} - {{ ticket.navette.horaireDepart|date('d/m/Y H:i') }}{% endif %}</td>
-								<td class='text-right'>
-									<a class='btn btn-info btn-xs' href='{{ path('sdf_billetterie_ticket_edit', { 'id' : ticket.id }) }}' role='button'>Options</a>
-									<a class='btn btn-primary btn-xs' href='{{ path('sdf_billetterie_accessBillet', { 'id' : ticket.id }) }}' role='button'>Imprimer &raquo;</a>
-								</td>
-							</tr>
-						{% endfor %}
-					</tbody>
-				</table>
-			</div>
-		</div>
-	{% endif %}
-
-	{% if availablePrices|length == 0 %}
-		<h2>Aucun billet disponible à l'achat.</h2>
-	{% else %}
-		<!-- Liste des billets disponibles -->
-		<div class='panel panel-default'>
-		<!-- Default panel contents -->
-			<div class='panel-heading'>
-				<h3>Billets disponibles</h3>
-			</div>
-			<div class='panel-body'>
-				<!-- Table -->
-				<table class='table'>
-					<thead>
-						<tr>
-							<th>Type de billet</th>
-							<th>Prix</th>
-							<th>Places restantes*</th>
-							<th class='text-right'>Actions</th>
-						</tr>
-					</thead>
-					<tbody>
-						{% for price in availablePrices %}
-							<tr>
-								<td>{{ price.nomTarif }}</td>
-								<td>{{ price.prix }} €</td>
-								<td>{{ remainingPlacesByPrice[price.id] }}</td>
-								<td class='text-right'>
-									<a class='btn btn-primary btn-xs' href='{{ path('sdf_billetterie_achatBillet', { 'typeBillet' : price.id }) }}' role='button'>Acheter &raquo;</a>
-								</td>
-							</tr>
-						{% endfor %}
-					</tbody>
-				</table>
-				<i>Le nombre de places restantes dépend de plusieurs paramètres, tels que la capacité de l'évènement, le nombre de places disponibles par tarif, le nombre de places disponibles par utilisateur.</i>
-				<strong>Il représente le nombre de places disponibles, pour vous, pour ce tarif. Pas nécessairement le nombre de places total disponible.</strong>
-			</div>
-		</div>
-	{% endif %}
-{% endblock %}
