--- src/SDF/BilletterieBundle/Resources/views/Pages/inscription.html.twig
@@ -1,61 +1,0 @@
-{% extends 'SDFBilletterieBundle::layout.html.twig' %}
-
-{% form_theme form 'bootstrap_3_layout.html.twig' %}
-
-{% block app_hero_content %}
-	<h1>Création d'un compte</h1>
-	<p>Veuillez compléter le formulaire ci-dessous.</p>
-{% endblock %}
-
-{% block app_content %}
-	<div class="container">
-		<div class="form_mdp">
-			<form action='{{ path('sdf_billetterie_subscribe_check') }}' method='{{ form.vars.method }}' {{ form_enctype(form) }}>
-				{{ form_widget(form) }}
-				<div class='form-group'>
-					<input class='btn btn-success' name='submit' type='submit' value='Inscription' />
-				</div>
-			</form>
-			{#
-			<form action="{{ path('sdf_billetterie_formulaireCreationCompte') }}" method="post">
-				<p><div class="input-group">
-					<span class="input-group-addon">Login</span>
-					<input type="text" id="username" name="username" class="form-control" aria-describedby="basic-addon1">
-				</div><br />
-
-								<div class="input-group">
-					<span class="input-group-addon">Prénom</span>
-					<input type="text" id="prenom" name="prenom" class="form-control" aria-describedby="basic-addon1">
-				</div><br />
-
-								<div class="input-group">
-					<span class="input-group-addon">Nom</span>
-					<input type="text" id="nom" name="nom" class="form-control" aria-describedby="basic-addon1">
-				</div><br />
-
-								<div class="input-group">
-					<span class="input-group-addon">Email</span>
-					<input type="email" id="mail" name="mail" class="form-control" aria-describedby="basic-addon1">
-				</div><br />
-
-								<div class="input-group">
-					<span class="input-group-addon">Mot de passe</span>
-					<input type="password" class="form-control" aria-describedby="basic-addon1"id="password" name="password" />
-				</div> <br />
-
-				<div class="input-group">
-					<span class="input-group-addon">Confirmation</span>
-					<input type="password" class="form-control" aria-describedby="basic-addon1"id="confirmation" name="confirmation" />
-				</div> <br />
-
-								<div class="input-group">
-					<span class="input-group-addon">Date de naissance</span>
-					<input type="date" class="form-control" placeholder="DD-MM-YYYY" aria-describedby="basic-addon1"id="birthday" name="birthday" />
-				</div> <br />
-				<input type="submit" value="Envoyer" />
-
-				</p>
-			</form> #}
-		</div>
-	</div>
-{% endblock %}
